Code Quality: fix the file name Duke to duke
