--- src/main/java/Duke/Launcher.java
@@ -1,17 +1,0 @@
-package Duke;
-
-import static javafx.application.Application.launch;
-
-/**
- * A launcher class to workaround classpath issues.
- */
-public class Launcher {
-
-    /**
-     * The main method for launch the Main class
-     * @param args
-     */
-    public static void main(String[] args) {
-        launch(Main.class, args);
-    }
-}
