+++ src/main/java/duke/Exceptions/DeleteWrongIndexException.java
@@ -1,0 +1,14 @@
+package duke.Exceptions;
+
+/**
+ * Represents DeleteWrongIndexException class
+ */
+public class DeleteWrongIndexException extends Exception {
+    /**
+     * The Constructor for DeleteWrongIndexException
+     * @param message
+     */
+    public DeleteWrongIndexException(String message) {
+        super(String.format("☹ OOPS!!! The target you delete is out of list, please enter again"));
+    }
+}
