+++ src/main/java/duke/Exceptions/NoCommandException.java
@@ -1,0 +1,15 @@
+package duke.Exceptions;
+
+/**
+ * Represents NoCommandException class
+ */
+public class NoCommandException extends Exception {
+
+    /**
+     * The Constructor for NoCommandException
+     * @param message
+     */
+    public NoCommandException(String message) {
+        super(String.format("☹ OOPS!!! I'm sorry, but I don't know what that means :-("));
+    }
+}
